refactor mapItem on CollectionViewer
