+++ src/app/postman/[collections]/types.ts
@@ -1,0 +1,17 @@
+import { HTTP_METHOD } from 'next/dist/server/web/http';
+
+export interface EssentialPostmanAPIResponse {
+  info: string;
+  item: string;
+  auth?: any;
+  globals: any[];
+  variable: any[];
+}
+
+export interface PostmanItemsApi {
+  name: string;
+  item?: PostmanItemsApi[];
+  request?: {
+    method: HTTP_METHOD;
+  };
+}
